REPORT.
I've made the following changes to my Flat class: translated the documentation lines into English, in functions added a description of what they does and their parameters.
I changed the design in Pandas, for example, used a link to a file, placed the stages of the work at the end of the project, added conclusions. I changed the histograms, got a graphical representation of the characteristics of the apartments.