Report
I've changed my jupiter notebook: I've downloaded and used data from the property.db database for analysis. I've made an analysis of the characteristics of the flats. In the OOP part, I've renamed the classDatabase file to models and added the remaining objects from the database to it. The executing part of classDatabase.py I've moved to the main.py. I've removed unnecessary items from the repository.